moving resources types back to the router
