+++ src/resources/controllers/resource-store/utils/get-resources-for-next-location/index.ts
@@ -1,5 +1,8 @@
+import {
+  ResourceStoreContext,
+  RouteResource,
+} from '../../../../../common/types';
 import { RouterContext } from '../../../../../index';
-import { ResourceStoreContext, RouteResource } from '../../../../common/types';
 import {
   getResourceIdentifier,
   getResourceIdentifiers,
