+++ src/resources/controllers/resource-store/utils/lru-cache/index.ts
@@ -3,7 +3,7 @@ import { StoreActionApi } from 'react-sweet-state';
 import {
   RouteResource,
   RouteResourceDataForType,
-} from '../../../../common/types';
+} from '../../../../../common/types';
 import { State } from '../../types';
 import { deleteResourceState } from '../manage-resource-state';
