+++ src/resources/controllers/resource-store/utils/manage-resource-state/index.ts
@@ -4,7 +4,7 @@ import {
   ResourceType,
   ResourceKey,
   RouteResourceResponse,
-} from '../../../../common/types';
+} from '../../../../../common/types';
 import { PrefetchSlice, State } from '../../types';
 
 export const getPrefetchSlice =
